% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_aug.R
\name{code_temperature_augmented}
\alias{code_temperature_augmented}
\title{Augmented a Stan file with alternative model blocks}
\usage{
code_temperature_augmented(stan_file = NULL)
}
\arguments{
\item{stan_file}{The path to the \code{Stan} program to use (\pkg{rstan}
package). It should be a character string file name.}
}
\value{
The new \code{Stan} file path.
}
\description{
Convenience function for decorating a Stan file with alternative model
blocks and construct a geometric bridge in between.
}
\details{
The function constructs a geometric bridge in between two posterior
densities that is proportional to the joint posterior in two model blocks.
A new \code{Stan} file will be created.
}
\examples{
\dontrun{
file_new=code_temperature_augment(stan_file="example/cauchy_mixture.stan")
}
}
