% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_aug.R
\name{code_tempeture_augmented}
\alias{code_tempeture_augmented}
\title{Augmented a Stan file with alternative model blocks}
\usage{
code_tempeture_augmented(stan_file = NULL)
}
\arguments{
\item{stan_file}{The path to the \code{Stan} program to use (\pkg{rstan}
package). It should be a character string file name.}
}
\value{
The new \code{Stan} file path.
}
\description{
Convenience function for decroating a Stan file with alternative model
blocks and contruct a geometric bridge in between.
}
\details{
The function constructs a geometric bridge in between two posterior
densities that is propotional to the joint posterior in two model blocks.
}
\examples{
\dontrun{
file_new=code_tempeture_augment(stan_file="example/cauchy_mixture.stan")
}
}
