% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lp.R
\name{extract_lp}
\alias{extract_lp}
\title{Extract path gradient from the augmented stan sampling results}
\usage{
extract_lp(output_file = "stan-output_path.txt")
}
\arguments{
\item{output_file}{The name of the output file that saves local variables, which
otherwise would be saved in stan. The default is \code{stan-output_path.txt}.}
}
\value{
The pointwise log density difference between two model blocks.
}
\description{
Extract path gradient from the augmented stan sampling results
}
\details{
Currently Stan cannot save local variables, while the path sampling gradient
often depends on partial sum of the target lp. This is an convenient function
that prints the need value during each leapfrog step and collect them afterwards.
}
