% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{g_lambda}
\alias{g_lambda}
\title{The link function}
\usage{
g_lambda(a, a_lower = 0.1, a_upper = 0.8)
}
\arguments{
\item{a}{transform temperature in [0,2].}

\item{a_lower}{The lower bound of the quadrature. When a < a_lower, the inverse
temperature  is 0 and the sample is from the base.}

\item{a_upper}{The upper bound of the quadrature. When a > a_upper, the inverse
temperature  is 1 and the sample is from the target.}
}
\value{
temperature lambda in [0,1].
}
\description{
transform temperature lambda to a.
}
