% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfit.R
\name{path_fit}
\alias{path_fit}
\title{Fit one adaptation of path sampling}
\usage{
path_fit(sampling_model, data_list = list(), a_lower, a_upper, b,
  K_logit, mu_logit, sigma_logit, K_gaussian, mu_gaussian, sigma_gaussian,
  all_a = NULL, all_log_lik = NULL, N_grid = 100, iter = 2000,
  max_treedepth = 8, thin = 2, visualize_progress = FALSE,
  output_file = "stan-output_path.txt")
}
\arguments{
\item{sampling_model}{The stan model generated from \code{\link{code_temperature_augmented}}.}

\item{data_list}{The list of data used in the original stan model.}

\item{a_lower}{The lower bound of the quadrature. When a < a_lower, the inverse
temperature  is 0 and the sample is from the base.}

\item{a_upper}{The upper bound of the quadrature. When a > a_upper, the inverse
temperature  is 1 and the sample is from the target.}

\item{b}{The current parameterization of the pseudo prior.}

\item{K_logit}{An integer, the length of the logit kernels.}

\item{mu_logit}{The location of logit kernels.}

\item{sigma_logit}{The width of logit kernels.}

\item{K_gaussian}{An integer, the length of the Gaussian kernels.}

\item{mu_gaussian}{The location of Gaussian kernels.}

\item{sigma_gaussian}{The width of Gaussian kernels.}

\item{all_a}{(optional) All sampled transformed tempeture from previous adaptations.
If not NULL, they will be remixed and used for normalization constant estimation.}

\item{all_log_lik}{(optional) All sampled pointwise log density difference between two
models from previous adaptations.  If not NULL, they will be remixed and used for
 normalization constant estimation.}

\item{N_grid}{The number of internal interpolations in parametric estimation
The default is 100.}

\item{iter}{The number of iterations for the augmented stan program.}

\item{max_treedepth}{The max tree depth  for the augmented stan program. A smaller
max_treedepth may increase exploration. efficiency in early adaptations.}

\item{thin}{The number of thining for the augmented stan program.}

\item{visualize_progress}{whether to visualize the progress.}

\item{output_file}{The name of the output file that saves local variables, which
otherwise would be saved in stan. The default is \code{stan-output_path.txt}.}
}
\value{
Names list of fit_main: the stan fit object; b: the updated pseudo prior
that approximate the marginal density after this adaptations;  all_a: all sampled
inverse temperature including the current adaptations; all_log_lik: all saved
log likelihood, including the current adaptations; path_post_z: the estimation of the
log normalization constants, path_post_a: the location of a in path_post_z.
}
\description{
Run path sampling with given parameters and pseudo priors.
}
\details{
The function fits one adaptations of path sampling with given
parameters and  pseudo priors that are specified by \code{b}. It runs the stan program,
run path sampling to compute the log normalization constants and log marginal,
and these two estimations with extra parametric regularization.
}
\examples{
\dontrun{
#' Make sure to pre-compile the stan optimization file
cat("Compiling Stan optimizer...")
update_model <- stan_model("solve_tempering.stan")
cat("done.\\n")
}


}
\seealso{
\code{\link{path_gradients}}, \code{\link{path_quadrature}},\code{\link{path_sample}}
}
