% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.R
\name{path_quadrature}
\alias{path_quadrature}
\title{Numerical quadrature using trapezoid rule}
\usage{
path_quadrature(a, u, a_lower = 0.1, a_upper = 0.8)
}
\arguments{
\item{a}{a vector, all sampled transformed parameters.}

\item{u}{a vector, pointwise gradient at those points. Often returned by
\code{\link{path_gradients}}.}

\item{a_lower}{The lower bound of the quadrature. When a < a_lower, the inverse
temperature is 0 and the sample is from the base.}

\item{a_upper}{The upper bound of the quadrature. When a > a_upper, the inverse
temperature is 1 and the sample is from the target.}
}
\value{
The pointwise  log normalization constant at all a.
}
\description{
Use numerical quadrature using trapezoid rule to compute the log normalization
constant from its potinwise gradient.
}
\details{
The function use numerical quadrature using trapezoid rule to compute
the normalization constant from its pointwise gradient \code{u}.
}
\references{
Gelman and Meng (1998). Simulating Normalizing Constants: From Importance Sampling to Bridge Sampling to Path Sampling. \emph{Statistical Science} \strong{13},
163-185.
}
