% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathtemp-package.R
\docType{package}
\name{pathtemp-package}
\alias{pathtemp-package}
\title{Adaptive Path Sampling and Continuous Tempering}
\description{
iteratively reduces the gap between the proposal and the target density, and provide a reliable
 normalizing constant  estimation with practical diagnostic using importance sampling theory. When
 equip simulated tempering with a continuous temperature, path tempering enables efficient sampling
 from multimodal densities.
}
\references{
Yuling Yao, Collin  Cademartori,Aki Vehtari,
Andrew Gelman Adaptive Path Sampling in Metastable Posterior Distributions
}
